refactor: decompose admin links manager into hooks and sub-components
